+++ derive/src/context/std/ops.rs
@@ -163,12 +163,18 @@ impl Context {
     ) -> proc_macro2::TokenStream {
         let Context {
             modulus,
+            cartridge,
             std,
+            num_bigint,
+            num_rational,
+            modtype,
             struct_ident,
             field_ident,
+            other_fields,
             ..
         } = self;
-        let generics = self.with_features(&[feature]);
+
+        let generics = self.with_features(&[feature.clone()]);
         let (impl_generics, ty_generics, where_clause) = generics.split_for_impl();
 
         let (struct_update, struct_update_deref) = self.struct_update(
@@ -180,7 +186,7 @@ impl Context {
             ],
         );
 
-        quote! {
+        let mut acc = quote! {
             impl
                 #impl_generics
             #std::ops::#trait_ident<#struct_ident#ty_generics> for #struct_ident#ty_generics
@@ -232,28 +238,204 @@ impl Context {
                     #struct_update_deref
                 }
             }
+        };
+
+        let generics = self.with_features(&[feature.clone(), parse_quote!(FlexibleRhs)]);
+        let (_, _, where_clause) = generics.split_for_impl();
+
+        let generics = self.with_features(&[
+            feature.clone(),
+            parse_quote!(PartialSubtraction),
+            parse_quote!(FlexibleRhs),
+        ]);
+        let (_, _, where_clause_sub) = generics.split_for_impl();
+
+        let generics = self.with_features(&[
+            feature.clone(),
+            parse_quote!(AssumePrimeModulus),
+            parse_quote!(PartialDivision),
+            parse_quote!(FlexibleRhs),
+        ]);
+        let (_, _, where_clause_div) = generics.split_for_impl();
+
+        let generics = self.with_features(&[
+            feature.clone(),
+            parse_quote!(AssumePrimeModulus),
+            parse_quote!(PartialSubtraction),
+            parse_quote!(PartialDivision),
+            parse_quote!(FlexibleRhs),
+        ]);
+        let (_, _, where_clause_sub_div) = generics.split_for_impl();
+
+        let generics = self.with_features(&[
+            feature.clone(),
+            parse_quote!(AssumePrimeModulus),
+            parse_quote!(PartialSubtraction),
+            parse_quote!(PartialMultiplication),
+            parse_quote!(PartialDivision),
+            parse_quote!(FlexibleRhs),
+        ]);
+        let (_, _, where_clause_sub_mul_div) = generics.split_for_impl();
+
+        for (rhs_ty, convert, where_clause) in &[
+            (quote!(u8), quote!(from_u8), where_clause),
+            (quote!(u16), quote!(from_u16), where_clause),
+            (quote!(u32), quote!(from_u32), where_clause),
+            (quote!(u64), quote!(from_u64), where_clause),
+            (quote!(u128), quote!(from_u128), where_clause),
+            (quote!(usize), quote!(from_usize), where_clause),
+            (quote!(i8), quote!(from_i8), where_clause_sub),
+            (quote!(i16), quote!(from_i16), where_clause_sub),
+            (quote!(i32), quote!(from_i32), where_clause_sub),
+            (quote!(i64), quote!(from_i64), where_clause_sub),
+            (quote!(i128), quote!(from_i128), where_clause_sub),
+            (quote!(isize), quote!(from_isize), where_clause_sub),
+            (
+                quote!(f32),
+                quote!(from_float_prim),
+                where_clause_sub_mul_div,
+            ),
+            (
+                quote!(f64),
+                quote!(from_float_prim),
+                where_clause_sub_mul_div,
+            ),
+            (
+                quote!(#num_bigint::BigUint),
+                quote!(from_biguint),
+                where_clause,
+            ),
+            (
+                quote!(#num_bigint::BigInt),
+                quote!(from_bigint),
+                where_clause_sub,
+            ),
+            (
+                quote!(#num_rational::Ratio<#num_bigint::BigUint>),
+                quote!(from_biguint_ratio),
+                where_clause_div,
+            ),
+            (
+                quote!(#num_rational::Ratio<#num_bigint::BigInt>),
+                quote!(from_bigint_ratio),
+                where_clause_sub_div,
+            ),
+        ] {
+            let value_rhs_move = quote! {
+                <#cartridge as #modtype::Cartridge>::#method(
+                    self.#field_ident,
+                    <#cartridge as #modtype::Cartridge>::#convert(rhs, #modulus),
+                    #modulus,
+                )
+            };
+            let value_rhs_clone = quote! {
+                <#cartridge as #modtype::Cartridge>::#method(
+                    self.#field_ident,
+                    <#cartridge as #modtype::Cartridge>::#convert(
+                        <#rhs_ty as #std::clone::Clone>::clone(rhs),
+                        #modulus
+                    ),
+                    #modulus,
+                )
+            };
+
+            let (update_move_move, update_move_clone, update_deref_move, update_deref_clone) =
+                if other_fields.is_empty() {
+                    (
+                        quote!(#struct_ident { #field_ident: #value_rhs_move }),
+                        quote!(#struct_ident { #field_ident: #value_rhs_clone }),
+                        quote!(#struct_ident { #field_ident: #value_rhs_move }),
+                        quote!(#struct_ident { #field_ident: #value_rhs_clone }),
+                    )
+                } else {
+                    (
+                        quote!(#struct_ident { #field_ident: #value_rhs_move, ..self }),
+                        quote!(#struct_ident { #field_ident: #value_rhs_clone, ..self }),
+                        quote!(#struct_ident { #field_ident: #value_rhs_move, ..*self }),
+                        quote!(#struct_ident { #field_ident: #value_rhs_clone, ..*self }),
+                    )
+                };
+
+            acc.extend(quote! {
+                impl #impl_generics #std::ops::#trait_ident<#rhs_ty> for #struct_ident#ty_generics
+                #where_clause
+                {
+                    type Output = #struct_ident#ty_generics;
+
+                    #[inline]
+                    fn #method(self, rhs: #rhs_ty) -> #struct_ident#ty_generics {
+                        #update_move_move
+                    }
+                }
+
+                impl
+                    #impl_generics
+                #std::ops::#trait_ident<&'_ #rhs_ty> for #struct_ident#ty_generics
+                #where_clause
+                {
+                    type Output = #struct_ident#ty_generics;
+
+                    #[inline]
+                    fn #method(self, rhs: &'_ #rhs_ty) -> #struct_ident#ty_generics {
+                        #update_move_clone
+                    }
+                }
+
+                impl
+                    #impl_generics
+                #std::ops::#trait_ident<#rhs_ty> for &'_ #struct_ident#ty_generics
+                #where_clause
+                {
+                    type Output = #struct_ident#ty_generics;
+
+                    #[inline]
+                    fn #method(self, rhs: #rhs_ty) -> #struct_ident#ty_generics {
+                        #update_deref_move
+                    }
+                }
+
+                impl
+                    #impl_generics
+                #std::ops::#trait_ident<&'_ #rhs_ty> for &'_ #struct_ident#ty_generics
+                #where_clause
+                {
+                    type Output = #struct_ident#ty_generics;
+
+                    #[inline]
+                    fn #method(self, rhs: &'_ #rhs_ty) -> #struct_ident#ty_generics {
+                        #update_deref_clone
+                    }
+                }
+            });
         }
+
+        acc
     }
 
     fn derive_bin_assign(
         &self,
         trait_ident: Ident,
         fn_ident: Ident,
-        impl_method: Ident,
+        cartridge_method: Ident,
         feature: Ident,
     ) -> proc_macro2::TokenStream {
         let Context {
             modulus,
+            cartridge,
             std,
+            num_bigint,
+            num_rational,
+            modtype,
             struct_ident,
             field_ident,
+            other_fields,
             ..
         } = &self;
-        let generics = self.with_features(&[feature]);
+        let generics = self.with_features(&[feature.clone()]);
         let (impl_generics, ty_generics, where_clause) = generics.split_for_impl();
 
         let (_, update) = self.struct_update(
-            impl_method,
+            cartridge_method.clone(),
             &[
                 parse_quote!(self.#field_ident),
                 parse_quote!(rhs.#field_ident),
@@ -261,7 +443,7 @@ impl Context {
             ],
         );
 
-        quote! {
+        let mut acc = quote! {
             impl#impl_generics #std::ops::#trait_ident<Self> for #struct_ident#ty_generics
             #where_clause
             {
@@ -279,6 +461,144 @@ impl Context {
                     *self = #update;
                 }
             }
+        };
+
+        let generics = self.with_features(&[feature.clone(), parse_quote!(FlexibleRhs)]);
+        let (_, _, where_clause) = generics.split_for_impl();
+
+        let generics = self.with_features(&[
+            feature.clone(),
+            parse_quote!(PartialSubtraction),
+            parse_quote!(FlexibleRhs),
+        ]);
+        let (_, _, where_clause_sub) = generics.split_for_impl();
+
+        let generics = self.with_features(&[
+            feature.clone(),
+            parse_quote!(AssumePrimeModulus),
+            parse_quote!(PartialDivision),
+            parse_quote!(FlexibleRhs),
+        ]);
+        let (_, _, where_clause_div) = generics.split_for_impl();
+
+        let generics = self.with_features(&[
+            feature.clone(),
+            parse_quote!(AssumePrimeModulus),
+            parse_quote!(PartialSubtraction),
+            parse_quote!(PartialDivision),
+            parse_quote!(FlexibleRhs),
+        ]);
+        let (_, _, where_clause_sub_div) = generics.split_for_impl();
+
+        let generics = self.with_features(&[
+            feature.clone(),
+            parse_quote!(AssumePrimeModulus),
+            parse_quote!(PartialSubtraction),
+            parse_quote!(PartialMultiplication),
+            parse_quote!(PartialDivision),
+            parse_quote!(FlexibleRhs),
+        ]);
+        let (_, _, where_clause_sub_mul_div) = generics.split_for_impl();
+
+        for (rhs_ty, convert, where_clause) in &[
+            (quote!(u8), quote!(from_u8), where_clause),
+            (quote!(u16), quote!(from_u16), where_clause),
+            (quote!(u32), quote!(from_u32), where_clause),
+            (quote!(u64), quote!(from_u64), where_clause),
+            (quote!(u128), quote!(from_u128), where_clause),
+            (quote!(usize), quote!(from_usize), where_clause),
+            (quote!(i8), quote!(from_i8), where_clause_sub),
+            (quote!(i16), quote!(from_i16), where_clause_sub),
+            (quote!(i32), quote!(from_i32), where_clause_sub),
+            (quote!(i64), quote!(from_i64), where_clause_sub),
+            (quote!(i128), quote!(from_i128), where_clause_sub),
+            (quote!(isize), quote!(from_isize), where_clause_sub),
+            (
+                quote!(f32),
+                quote!(from_float_prim),
+                where_clause_sub_mul_div,
+            ),
+            (
+                quote!(f64),
+                quote!(from_float_prim),
+                where_clause_sub_mul_div,
+            ),
+            (
+                quote!(#num_bigint::BigUint),
+                quote!(from_biguint),
+                where_clause,
+            ),
+            (
+                quote!(#num_bigint::BigInt),
+                quote!(from_bigint),
+                where_clause_sub,
+            ),
+            (
+                quote!(#num_rational::Ratio<#num_bigint::BigUint>),
+                quote!(from_biguint_ratio),
+                where_clause_div,
+            ),
+            (
+                quote!(#num_rational::Ratio<#num_bigint::BigInt>),
+                quote!(from_bigint_ratio),
+                where_clause_sub_div,
+            ),
+        ] {
+            let value_rhs_move = quote! {
+                <#cartridge as #modtype::Cartridge>::#cartridge_method(
+                    self.#field_ident,
+                    <#cartridge as #modtype::Cartridge>::#convert(rhs, #modulus),
+                    #modulus,
+                )
+            };
+            let value_rhs_clone = quote! {
+                <#cartridge as #modtype::Cartridge>::#cartridge_method(
+                    self.#field_ident,
+                    <#cartridge as #modtype::Cartridge>::#convert(
+                        <#rhs_ty as #std::clone::Clone>::clone(rhs),
+                        #modulus
+                    ),
+                    #modulus,
+                )
+            };
+
+            let (update_rhs_move, update_rhs_clone) = if other_fields.is_empty() {
+                (
+                    quote!(#struct_ident { #field_ident: #value_rhs_move }),
+                    quote!(#struct_ident { #field_ident: #value_rhs_clone }),
+                )
+            } else {
+                (
+                    quote!(#struct_ident { #field_ident: #value_rhs_move, ..*self }),
+                    quote!(#struct_ident { #field_ident: #value_rhs_clone, ..*self }),
+                )
+            };
+
+            acc.extend(quote! {
+                impl
+                    #impl_generics
+                #std::ops::#trait_ident<#rhs_ty> for #struct_ident#ty_generics
+                #where_clause
+                {
+                    #[inline]
+                    fn #fn_ident(&mut self, rhs: #rhs_ty) {
+                        *self = #update_rhs_move;
+                    }
+                }
+
+                impl
+                    #impl_generics
+                #std::ops::#trait_ident<&'_ #rhs_ty> for #struct_ident#ty_generics
+                #where_clause
+                {
+                    #[inline]
+                    fn #fn_ident(&mut self, rhs: &'_ #rhs_ty) {
+                        *self = #update_rhs_clone;
+                    }
+                }
+            });
         }
+
+        acc
     }
 }